# internal/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A self-hostable metadata API that is backward compatible with R——.",
        "title": "bookinfo.pro api",
        "contact": {
            "url": "https://github.com/blampe/rreading-glasses"
        },
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/author/{authorId}": {
            "get": {
                "description": "This returns an extremely \"fat,\" un-paginated payload -- in many cases many megabytes in size. The design of this endpoint was at the core of R——'s performance problems.",
                "summary": "Fetch author metadata by foreign ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Return the author with only this edition loaded; more performant",
                        "name": "editionId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.AuthorResource"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Refresh an author",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to refresh all of the author's works and editions as well",
                        "name": "full",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/book/asin/{asin}": {
            "get": {
                "description": "Returns an ID appropriate for /book/. This lookup might fail if the server hasn't already loaded the edition.",
                "summary": "Look up a foreign edition ID by ASIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ASIN to look up",
                        "name": "asin",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            }
        },
        "/book/isbn/{isbn}": {
            "get": {
                "description": "Returns an ID appropriate for /book/. This lookup might fail if the server hasn't already loaded the edition.",
                "summary": "Look up a foreign edition ID by ISBN10 or ISBN13",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ISBN to look up",
                        "name": "isbn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            }
        },
        "/book/{editionID}": {
            "get": {
                "description": "Fetch a book (edition) by foreign ID. Confusingly an edition's metadata is the same format as a work's metadata.",
                "summary": "Fetch an edition of a work",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Edition ID",
                        "name": "editionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.workResource"
                        }
                    }
                }
            }
        },
...
        "/recommended": {
            "get": {
                "description": "This matches what you see on the upstream website's \"trending\" page.",
                "summary": "Fetch rending work IDs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The page of results; not supported by G—R—",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.RecommentationsResource"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Search both authors and works for the given query.",
                "summary": "Perform a freetext search query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the query string",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.SearchResource"
                            }
                        }
                    }
                }
            }
        },
        "/search/batch": {
            "post": {
                "description": "Search both authors and works for multiple queries at once. All queries are processed concurrently, allowing the underlying batched GraphQL client (1-second window, batch size 25) to combine multiple Search operations into a single HTTP request to Hardcover. This significantly reduces API calls to Hardcover's rate-limited API (60 requests/minute). For example: 10 sequential /search calls = 10 API requests, but 1 /search/batch call with 10 queries = 1 API request (if within batch window).",
                "summary": "Perform multiple freetext search queries in a single request",
                "parameters": [
                    {
                        "description": "array of query strings",
                        "name": "queries",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.BatchSearchResource"
                        }
                    }
                }
            }
        },
        "/series/{seriesId}": {
            "get": {
                "summary": "Fetch series metadata by foreign ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Series ID",
                        "name": "seriesId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.SeriesResource"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/work/{workId}": {
            "get": {
                "description": "Confusingly a work's metadata is actually just the \"best\" edition's metadata.",
                "summary": "Returns work metadata by foreign ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "workId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.workResource"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal.AuthorResource": {
            "type": "object",
            "properties": {
                "AverageRating": {
                    "type": "number"
                },
                "Description": {
                    "type": "string"
                },
                "ForeignId": {
                    "type": "integer"
                },
                "ImageUrl": {
                    "type": "string"
                },
                "KCA": {
                    "description": "New fields.",
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "RatingCount": {
                    "type": "integer"
                },
                "Series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.SeriesResource"
                    }
                },
                "Url": {
                    "type": "string"
                },
                "Works": {
                    "description": "Relations.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.workResource"
                    }
                }
            }
        },
        "internal.BatchSearchResource": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/internal.SearchResource"
                        }
                    }
                }
            }
        },
        "internal.RecommentationsResource": {
            "type": "object",
            "properties": {
                "workIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "internal.SearchResource": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/internal.SearchResourceAuthor"
                },
                "bookId": {
                    "type": "integer"
                },
                "workId": {
                    "type": "integer"
                }
            }
        },
        "internal.SearchResourceAuthor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal.SeriesResource": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "ForeignId": {
                    "type": "integer"
                },
                "KCA": {
                    "description": "New fields",
                    "type": "string"
                },
                "LinkItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.seriesWorkLinkResource"
                    }
                },
                "Title": {
                    "type": "string"
                }
            }
        },
...
    }
}